+++ libraries/TCA9554/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2025-01-12
+- fix **begin()** mode parameter (thanks to CrispyPyro)
+- update readme.md.
+- add derived classes PCA9554 and PCA9534
+
+
 ## [0.1.0] - 2025-01-09
 - initial version (based upon TCA9555)